[FIX] Coding style compliance
